+++ src/lib/pages/evm-contract-details/types.ts
@@ -9,11 +9,18 @@ export enum TabIndex {
   Transactions = "transactions",
 }
 
+export enum InteractTabsIndex {
+  Read = "read",
+  Write = "write",
+  ReadProxy = "read-proxy",
+  WriteProxy = "write-proxy",
+}
+
 export enum EvmContractDetailsContractTabs {
-  Code = "code",
-  Compiler = "compiler",
-  Abi = "abi",
-  ByteCode = "byteCode",
+  Code = "Code",
+  Compiler = "Compiler",
+  Abi = "ABI",
+  ByteCode = "ByteCode",
 }
 
 export const zEvmContractDetailsQueryParams = z.object({
@@ -25,6 +32,14 @@ export const zEvmContractDetailsQueryParams = z.object({
       .optional()
       .transform(() => TabIndex.Overview),
   ]),
+  selectedType: z.union([
+    z.nativeEnum(InteractTabsIndex),
+    z
+      .string()
+      .optional()
+      .transform(() => InteractTabsIndex.Read),
+  ]),
+  selectedFn: z.string().optional(),
 });
 
 export enum TxsTabIndex {